Refactored ConnectionLogger to base utility class
