+++ CoreUtilities/src/au/gov/asd/tac/constellation/utilities/rest/Response.java
@@ -72,8 +72,7 @@ public abstract class Response {
 
     protected Response(final int code, final String message, final Map<String, List<String>> headers, final byte[] bytes) throws IOException {
         this(code, message, headers, bytes, true);        
-        //ConnectionLogger.log(Level.FINE, toString(), null);
-        
+        ConnectionLogger.getInstance().log(Level.FINE, toString(), null);
     }
 
     protected Response(final int code, final String message, final Map<String, List<String>> headers, final byte[] bytes, final boolean isJson) throws IOException {
